Main app renamed
