--- src/server/main/fixtures/initial.json
@@ -1,42 +1,0 @@
-[
-    {
-        "model": "main.portionitem",
-        "pk": 1,
-        "fields": {
-            "name": "Carbohydrates",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.portionitem",
-        "pk": 2,
-        "fields": {
-            "name": "Fruits & Vegetables",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.portionitem",
-        "pk": 3,
-        "fields": {
-            "name": "Protein",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.portionitem",
-        "pk": 4,
-        "fields": {
-            "name": "Dairy",
-            "is_default": true
-        }
-    },
-    {
-        "model": "main.portionitem",
-        "pk": 5,
-        "fields": {
-            "name": "Oils & Fats",
-            "is_default": true
-        }
-    }
-]
