+++ src/server/mate/admin.py
@@ -15,7 +15,7 @@ class LogItemInline(admin.TabularInline):
 
 
 @admin.register(User)
-class MainUserAdmin(UserAdmin):
+class UserAdmin(UserAdmin):
     fieldsets = (
         (None, {"fields": ("email", "password")}),
         (_("Personal info"), {"fields": ("first_name", "last_name", "picture")}),
