+++ src/server/mate/apps.py
@@ -1,9 +1,9 @@
 from django.apps import AppConfig
 
 
-class MainConfig(AppConfig):
+class MateConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "main"
+    name = "mate"
 
     def ready(self):
-        import main.signals  # noqa
+        import mate.signals  # noqa
