+++ src/server/rest/serializers.py
@@ -4,7 +4,7 @@ from django.core.exceptions import ValidationError
 from django.contrib.auth.password_validation import validate_password
 from rest_framework import serializers
 
-from main import models
+from mate import models
 from auth.models import User
