+++ src/server/server/urls.py
@@ -4,7 +4,7 @@ from django.conf import settings
 from django.conf.urls.static import static
 
 urlpatterns = [
-    path("", include("main.urls")),
+    path("", include("mate.urls")),
     path("api/", include("rest.urls")),
     path("admin/", admin.site.urls),
     path("graphql", include("graph.urls")),
